src/python-producer/requirements.txt
kafka-python>=2.0.1,<2.1
lxml>=4.5.2,<4.6
